Describe the regions of the cube that we care about.
The six face points are already set.

Since we can only put edge pieces on edges and corner pieces on corners, we have a region for edges and a region for corners.

We include the point in the middle of the cube to keep the numbers simpler, but we just set it to 0 and don't put it in any regions.